Move TypeRef and Primitive into respective folders
