+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/mod.rs
@@ -5,11 +5,13 @@ use crate::ts_ast::{source_map::SourceMapped, TsAst};
 
 pub mod func;
 pub mod opt;
+pub mod primitive;
 pub mod record;
 pub mod service;
 pub mod tuple;
 pub mod type_alias;
 pub mod type_param;
+pub mod type_ref;
 pub mod variant;
 pub mod vec;
